Completed JWT auth and authMiddleware to protect user and goal routes

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-Handler");
 const { findByIdAndUpdate } = require("../models/goalModel");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 //@desc GET all Goals
 //@route  GET /api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -21,6 +22,7 @@ const createGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(201).json(goal);
 });
@@ -33,6 +35,18 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal Not Found");
+  }
+  const user = await User.findById(req.user.id);
+
+  // Check for user ID
+  if (!user) {
+    res.status(401);
+    throw new Error("User not Found");
+  }
+  // Make sure the goal user matches the logged in user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -50,6 +64,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal Not Found");
+  }
+  const user = await User.findById(req.user.id);
+
+  // Check for user ID
+  if (!user) {
+    res.status(401);
+    throw new Error("User not Found");
+  }
+  // Make sure the goal user matches the logged in user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authroized");
   }
   await goal.deleteOne();
   res.status(200).json({ id: req.params.id });
